backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
